Refactor Vulkan renderer for improved type safety and SSAO enhancements
- Added new cvar `r_ssaoDebugView` for SSAO debug visualization options.
- Updated Vulkan renderer functions to cast loop indices and sizes to appropriate types for better type safety.
- Introduced unused parameter suppression in various functions to clean up warnings.
- Enhanced shader module management by adding SSAO debug shaders for improved rendering diagnostics.
- Refactored several functions to use `const` qualifiers for better code clarity and safety.
- Improved handling of Vulkan command buffer sizes to prevent potential overflow issues.

+++ src/renderers/vulkanrenderer/vk_flares.c
@@ -232,7 +232,7 @@ void RB_AddDlightFlares( void ) {
 	if ( tr.world )
 		fog = tr.world->fogs;
 
-	for ( i = 0 ; i < backEnd.refdef.num_dlights; i++, l++ ) {
+	for ( i = 0 ; i < (int)backEnd.refdef.num_dlights; i++, l++ ) {
 
 		if ( fog )
 		{
